refactor(WarekiPicker): client専用処理をclient/に切り出す
'use client'が必要な実装をclient/ディレクトリに切り出す。
公開バレル(index.ts)自体は'use client'を持たないため、
Server Componentからも参照だけは可能になる。
