+++ packages/smarthr-ui/src/components/WarekiPicker/client/WarekiPicker.tsx
@@ -2,9 +2,9 @@
 
 import { type ComponentProps, type FC, useCallback } from 'react'
 
-import { useLatest } from '../../hooks/useLatest'
-import { useIntl } from '../../intl'
-import { DatePicker } from '../DatePicker'
+import { useLatest } from '../../../hooks/useLatest'
+import { useIntl } from '../../../intl'
+import { DatePicker } from '../../DatePicker'
 
 type Props = Omit<ComponentProps<typeof DatePicker>, 'showAlternative'>
